[Fix] 리팩터링 일부 진행

+++ favor/src/main/java/com/favor/favor/photo/PhotoService.java
@@ -20,12 +20,14 @@ import static com.favor.favor.exception.ExceptionCode.FILE_NOT_FOUND;
 import static com.favor.favor.exception.ExceptionCode.SERVER_ERROR;
 
 @Service
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class PhotoService {
     private  String bucketName = "favor-app-bucket";
     private final AmazonS3Client amazonS3Client;
 
     //S3에 사진 업로드
+    @Transactional
     public String uploadFileToS3(String fileName, MultipartFile file) {
 
         try {
@@ -43,8 +45,10 @@ public class PhotoService {
     }
 
     //S3 에서 사진 삭제
+    @Transactional
     public void deleteFileFromS3(String fileName) {
         try {
+            if (fileName == "") return;
             amazonS3Client.deleteObject(bucketName, fileName);
         } catch (SdkClientException e) {
             throw new CustomException(e, FILE_NOT_FOUND);
